Fix adding a notetype when using the name of a standard notetype
This fixes a bug where the user couldn't add a notetype with the name
of one of the standard notetypes(even if it didn't existed), because
the code checked the new name against the existing notetypes names +
the standard notetypes names so it behaved like the name already
existed.

The change makes the code look only at the existing notetypes names.