Fix background color update for TemplatePreviewerFragment in CardTemplateEditor
The fragment's view was accessed to set the background on it but it
wasn't available at that moment resulting in an IllegalStateException.

The fragment's view wasn't available because although the transaction
that adds the fragment is done synchronously it only takes the fragment
to the parent's CURRENT lifecycle state(still too early in the lifecycle).