Rephrase "Abort learning"
Those two strings appears as actions to which you can associate a
gesture or a shortcut.

I must admit that I've no idea what "Abort learning" mean, until I
looked at the code.

I hesitated to use "Close the Deck", because, after all, from the user
point of view, we were opening a deck. The name "reviewer" is rarely
used. But that's already the name used in our documentation and in
other settings string, so I decided to keep it.

Admittedly, I'm not a native English speaker, so it's hard for me to
know whether a native would understand directly what it means.