Ensure menu entries' titles are at most 28 char long
This reuse the code for preference's title.

Its distinct issues, because the description mentioned the limit, and
the description must be a constant.

This commit also updated all string resources to add the 28 character
limit.

It should be noted that, while most menus are shown as context
menu/overflow menu, there is one exception. The sidebar menu. Still,
it seems to me that 28 characters for statistics, card browser, deck
picker, is also a reasonable constraint. The side menu is relatively
small, and the text is big.
Admittedly, we allow for two lines there if necessary, so it should
still be okay with long-ish strings.