Remove the navigation drawer code from StudyOptionsActivity
I think this was some legacy code left behind. There's no reason to
have StudyOptionActivity inherit NavigationDrawerActivity because it
doesn't show the drawer(just the up arrow to go to DeckPicker, its parent)
and the drawer doesn't make sense anyway as this activity is a direct sub
activity of DeckPicker.