NF: Remove else -> when possible
When `when` is used on an enum, the `else ->` can always be
transformed into a complete enumeration of the enum's entries.

This ensure that each time a new entry is added, there is a warning
and the developer must decide whether or not a change is needed. While
the `else` may hide some required change.

So, for the sake of future developement, I removed `else ->` when it
makes sense.

I should note that I kept some `else ->` when the `when` consider only
a very small part of all possible enum value, in order to reduce the
size of the code.