Fix onBackPressed deprecation in StudyOptionsActivity
The previous code in onBackPressed() was just setting the result code
to RESULT_OK and then finishing the activity so I set that result code
directly in onCreate(), the associated fragment can change it as it
chooses.

With this change(vs. adding a OnBackPressedCallback) we get the nice
peek animation when dragging from the left to go back(
enableOnBackInvokedCallback has to be enabled for the application in
the manifest file).