feat: add ViewerAction and MenuDisplayType
ViewerAction is similar to ActionButtonStatus, but in an Enum

It also is similar to ViewerCommand and will replace it when gestures are implemented in the new reviewer

---

MenuDisplayType is a set of three display options for an action in the reviewer menu.

The legacy `If room` setting isn't being added because it is a setting for the *developers* to use while designing a menu. For a *user* it is bad, because:
- It isn't obvious how it works
- It only makes a difference when the app window is resized on bigger screens. In phones it won't make a difference, which represents most of our users.
- The difference of that resize is mostly 1-2 icons to be added or removed, which doesn't help much
- The reviewer ignores `screenSize` configuration changes, so all of that only apply if the reviewer is opened and closed

So I removed it to simplify things