pre-commit restore unstaged data
The documentation state that the function should return, but actually
it exists. This is repaired.

Removing `set -e` ensure the script does not stop if the linter fails.
This is probably what we expected given that we put the return value
in a variable.

This would ensure that the patch that was saved is always restored.

I don't expect this to cause issues because the remaining of the
changes are not destructive.

I manually tested that the bug can't be reproduced with this
script. However, I've not tested the change in case of `git apply`
failure, as I don't know how to force apply to fail.

Fixed: #17593