Deeplink to reviewer closes everything
Currently, the code ensured there was at most one IntentHandler. We
honestly don't care about it. The intended meaning was that there is a
single reviewer.

This bug was probably introduced in #17422 with commit
f26aa95bdf4bb0fb378da1bf626a70eecb46a2e5.

I would love to test to avoid another regression. However, it's really
not clear to me how to deal with shortcuts, widget or remainders as
they are outside of the main app.

I discovered that having a single reviewer is still not
enough. Because what is below the reviewer could be nonsense. Instead,
I followed standard practice for deep link. I close everything and
create a fresh stack.

Bug #16956.
Fixes #17664.

Co-authored-by: Mike Hardy <github@mikehardy.net>