Simplify CustomStudyDialog
I need to admit that Aditya13s in
https://github.com/ankidroid/Anki-Android/issues/17665#issuecomment-2563862406
has a point. The implementation of the inner functions such as
`customStudy` is slightly hard to understand.

The goal of the block, taking a custom study request's dsl and returning
nothing is not immediately clear, until you read the implementation of
the blocks that are sent to it and realize that this is just supposed
to create the request, except for the DID part.

It's not even clear when looking at the block why they are ignoring
the did.

Given that the only difference between the customStudy and the limit
extension different is only the listener's call, this can clearly be
simplified.

Also, I think it's nice to explain at the end why there is no reason
to save anything for the extensions. And to ensure that if there is an
extra option later, this `when` is not forgotten.