Rename model to note type in content provider
In this case, I only renamed private values and local variables,as the
content provider is used to provide content to external entities, its
API should not be changed. I also added some documentation to explain
why we have those old names.

We could consider duplicating "model" to "note type" in the entire
API, and noting "model" as deprecated. But I really don't think that's
it's worth deprecating anything. And if we don't plan to deprecate, I
don't think it's worth having duplicate API either.

Co-authored-by: David Allison <62114487+david-allison@users.noreply.github.com>