Fix several ActivityNotFoundExceptions
In certain screens of the app(like shared decks and account login) the
user could trigger an ACTION_VIEW from the content, which will
delegate to our custom tabs implementation. This will call the platform
launchUrl() method which would fail if there's no app to handle it.

Also added guards for the several extensions openUrl(), the check
AdaptionUtil.hasWebBrowser() is failing in a few cases.