Correct matchesJsonValue
`expectedMap` used to associated the keys of the `expectedValue` to the
values of `actualValue`.

As `differtKeys` also looked at the values in `actualValue`, it means
the values of `expectedValues` were always ignored.

I hope that this new writting, checking for the equality of the set of
keys, and then for the equalities of value, will be more readable and
hence improve the test suite.

Luckily enough we only had one failing test. I corrected the expected
value in the test. As we use the default filtered deck provided by the
back-end I don't expect the implementation to be buggy here.