NF: Comment DeckNode
I found DeckNode slightly hard to use, as it was not clear exactly
what each value was and what each function does.

For example, I would not have guessed that the parent of the top-level
deck is a root DeckNode that is not actually a deck. Or that the
counts included the cards to review in children decks.

I also modified an access to the parent to explain why `get()!!` is
safe.

Admittedly, I don't think that we should have used a WeakReference for
the parent here. The whole tree should have the same lifespan. If it's
not the case and a children outlive its parents, it's already an issue
and being able to release its parent may not save thas much memory overall.