Removing Q: and A: when a single side exists
Admittedly, for "answer XXX" this now lead the ambiguity of whether
this will have an impact on the question side, for people that set the
binding before the current update. I believe that this is still the
less surprising option

I also remove the string display_binding_card_side_both. While it was
simplifying slightly the code, it was at the cost of adding a
useless translation, I don't believe it was worth it.