Cause gesture to ask on which side it is executed
This ensure the behavior is consistent with other bindings.
More importantly for me, it'll allow to ensure that the "answer XXX"
gesture does not also trigger a flip, which is a problem I currently
have when I need to scroll to see the entire card and ends up flipping
the card by accident.