Limit size of image that can be used in the crop editor
If the uncompressed image is too big, the ImageView used by the crop
editor will fail when it will be drawn by the system.
The fix adds code that looks at the image to check against a set limit
and only then starts the normal crop loading URI code.

See #17378.