Remove "Close review screen and sync"
Many reason for this. First one being that the sync don't even occur (at least on my device). Also, that, according to Brayan this action is not very natural here.

I entirely remove the action. This way, if a user had it assigned to a shortcut, gesture, they will realize the reviewer stop closing, and if they thought the device was synced they'll finally know no sync will occur.

fixed #17586