feat(card-browser): add selection toggle in header
* A select all/deselect button was added
* Select all/none menu items were removed

This also makes '0 selected' a valid state, as it was
unintuitive that pressing 'select none' would change mode
if the 'back button' above did the same thing.

This state is only available through 'select none'
tapping on the last selected row disables multiselect

Otherwise, a user woudld need to long press a row,
then move to the top right of the screen to disable
multiselect, which felt frustrating to me.

-> modifies `endMultiSelectMode` logic
-> modifies `flowOfIsInMultiSelectMode`
-> Card Browser only shows 'preview' if no rows selected
-> last focus remains for Tablet Mode

Fixes 18572