Remove spaces of "Shake device" label (<= API 30)
On API 30 or below, Maracas emoji isn't shown on the "Shake device" label,
but the space for its replacement ("  ") and the space placed between the emoji and the string are still shown,
which causes unnatural blank or misalignment on the screen.

This PR removes the two kind of spaces on API 30 or below.