fix: sound deadlock when playing too fast
my guess is that the Dispatchers.IO thread pool gets full if too many sound tasks are opened quickly

if the thread pool is full, it isn't possible to launch another job to cancel the previous ones. That way, the scope gets deadlock

The issue is fixed by using a mutex, so the audio tasks are serialized, and only one audio can play at a time

It is also reproducible in the browser previewer.

it probably is harder to reproduce it on the old reviewer because the answer buttons have delays, and because the card rendering is slower, which means that it takes longer to parse the card data and start running the new audio files

Using Dispatchers.IO.limitedParallelism(1) would also work, but that would block the ttsPlayer initialization, so I went with a mutex.