Fix double toolbar bug in ManageSpaceActivity
The fix consists in making ManageSpaceActivity inherit SingleFragmentActivity,
this allows it to properly handle its content fragment ManageSpaceFragment
which is a SettingsFragment(designed to work within a SingleFragmentActivity).

After adding yet another hack, SingleFragmentActivity will now check to see if
the expect fragment is null(ManageSpaceActivity is started by the system) and
the inheriting class is ManageSpaceActivity. If it is, it will initialize with
the expected fragment(ManageSpaceFragment) or will revert to the original code.