Remove fragmented parameter from DeckPickerViewModel
DeckPicker was passing directly the fragmented status to its viewmodel
when creating it, but as the viewmodel can survive longer that the
activity, this could lead to cases were there is a disconnect between
the viewmodel and DeckPicker about the "fragmented" status(especially
as the viewmodel fragmented status was never updated for its entire
lifetime).

In certain cases(#19131) this could lead to crashes because if the
viewmodel is kept alive but the activity is recreated(like when
modifying the split mode and resizing), in DeckPicker.onCreate() we
would use the viewmodel incorrect fragmented status and look for
views not found in the actual(correct) layout we are inflating.

IMO this also fixes an architecture issue, because the viewmodel
shouldn't have/keep knowledge about how the UI is structured.