Make backend CardInfo page title to use the proper sentence case
getIntent() was changed to handle the sentence case for these strings
because the CardInfoDestination is used both in a "Context" class and
in a ViewModel where the context is not available(toSentence() requires
a context reference).

This target string inserts another string as context and we currently have
two combination so a sentence case string was added for each possible
combination.

The backend end string uses BiDi characters types FSI and PDI, see
bottom of the tabel at:

https://en.wikipedia.org/wiki/Bidirectional_text#Table_of_possible_BiDi_character_types