feat(reminders): add multiprofile ID to ReviewReminder
GSoC 2025: Review Reminders

As discussed on Discord, once the upcoming multiprofile system is implemented, we will need to fire review reminder notifications for all profiles, not just the active one. Currently however, once a notification is scheduled by AlarmManagerService and passed to the OS, there is no indication of which profile it is attached to as ReviewReminder does not include profile information. This is a problem because once the notification fires, it will need to open the corresponding collection to check if there are cards due. Without storing some sort of profile ID in ReviewReminder, the notification code will need to open each profile's collection one after the other to find the deck it is looking for.

This small change rectifies this issue by adding a profileID string field to ReviewReminder. The field is a string because most discussion around the profile ID in Discord and in candidate GSoC proposals revolves around using either a UUID or a string name. In the case of a UUID, we cannot use the built-in Java UUID type because it is not Serializable. Hence, it will need to be converted to a string before it is stored in ReviewReminder. In both cases, a string is sufficient.