Inline notetype name in NotetTypeFieldsEditor
Moves the notetype name from being the subtitle of the activity to being
shown in the activity layout, at the top. This makes the toolbar less
cluttered and makes it easier to handle the toolbar if this screen is
to be used in a multi pane(notetypes list panel + selected notetype fields)
configuration.

Also refactor the layout a bit.