Fix null deck view in CardBrowser multi select mode collector
If the SearchView was opened and the user selected a row(and entered
multiselect mode) the new deck name related views would be changed but
they weren't found in the layout(the SearchView is still visible at
that moment it seems).

The fix is to update the views in updateMultiselectMenu() which gets
triggered by an invalidate call after the search view is closed.