Fix text selection highlight color for outlined TextInputLayouts
Because we don't use the proper colors for colorOnPrimaryContainer the
selection color was white for this widgets and it was basically invisible
in light themes.

As a temporary fix the style for these widgets was changed to introduce a
theme overlay so we change the text selection color to a more suitable
value only for these widgets.